backend/requirements.txt
# WWMAA Backend - Python Dependencies
# Testing Infrastructure

# Core Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-xdist==3.5.0  # Parallel test execution

# Test Data Generation
faker==20.1.0

# HTTP Testing
httpx==0.25.2  # For testing async HTTP clients
respx==0.20.2  # Mock HTTP requests

# Web Framework (FastAPI/Flask)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0  # WebSocket support for real-time chat

# Database & Caching
redis==5.0.1
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External API Clients
stripe==7.7.0
requests==2.31.0
urllib3==2.1.0  # Required for connection pooling in requests

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# Scheduling (for dunning reminders)
APScheduler==3.10.4

# QR Code Generation (for event RSVPs)
qrcode[pil]==7.4.2
Pillow==10.1.0

# OpenAI Integration (for content indexing)
openai==1.6.1
tiktoken==0.5.2

# Profanity Filter (for chat moderation)
better-profanity==0.7.0

# Performance Monitoring (Prometheus)
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# OpenTelemetry Instrumentation (Sprint 7 - US-065)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0

# Error Tracking & Alerting (Sprint 7 - US-066)
sentry-sdk[fastapi]==1.40.0

# Security & Input Validation (Sprint 7 - US-070)
bleach==6.1.0  # HTML sanitization (XSS prevention)
python-magic==0.4.27  # File type detection (magic bytes verification)
dnspython==2.4.2  # DNS checks for email validation
pip-audit==2.6.1  # Security vulnerability scanning
